services/motif-service/requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
flatbuffers==23.5.26
humanfriendly==10.0
idna==3.4
imageio==2.32.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
lazy_loader==0.3
llvmlite==0.41.1
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.1
onnxruntime==1.16.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.0.1
platformdirs==4.0.0
pooch==1.8.0
protobuf==4.25.0
PyMatting==1.1.11
pyreadline3==3.4.1
referencing==0.30.2
rembg==2.0.52
requests==2.31.0
rpds-py==0.12.0
scikit-image==0.22.0
scipy==1.11.3
sympy==1.12
tifffile==2023.9.26
tqdm==4.66.1
urllib3==2.0.7
